requirements.txt
# ========== Core Framework ==========
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.40.0
pydantic>=2.0.0

# ========== Environment & Configuration ==========
python-dotenv>=1.0.0

# ========== Database ==========
pymongo>=4.10.0

# ========== LangChain Core ==========
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# ========== LangChain Integrations ==========
langchain-google-genai>=2.0.0    # Required for GoogleGenerativeAIEmbeddings
langchain-mongodb>=0.5.0          # Required for MongoDBAtlasVectorSearch

# ========== LangGraph ==========
langgraph>=0.2.0

# ========== Document Processing ==========
pypdf>=4.0.0                      # Required by PyPDFLoader

# ========== Frontend Dependencies ==========
requests>=2.31.0
Pillow>=10.0.0

# ========== Google AI ==========
google-generativeai>=0.8.0        # Required for ConnectLLM (Gemini API)
